add show_or_export tests
