+++ tests/conftest.py
@@ -1,0 +1,8 @@
+import pytest
+
+
+@pytest.fixture
+def plt_mock(mocker):
+    plt = mocker.patch("eddington_matplotlib.util.plt")
+    figure = plt.figure.return_value
+    return dict(plt=plt, figure=figure)
